// atividade-4/web-app/src/constants/types/Cruise.ts
export type Cruise = {
  id: number;
  nome_navio: string;
  nome_do_cruzeiro: string;
  destino: string;
  origem: string;
  porto_embarque: string;
  porto_desembarque: string;
  data_embarque: string;
  quantidade_noites: number;
  numero_cabines: number;
  lugares_visitados: string;
  valor_pessoa: number;
};